subreg: Add -fsplit-wide-types-early (PR88233)
Currently the second lower-subreg pass is run right before RA.  This
is much too late to be very useful.  At least for targets that do not
have RTL patterns for operations on multi-register modes it is a lot
better to split patterns earlier, before combine and all related
passes.

This adds an option -fsplit-wide-types-early that does that, and
enables it by default for rs6000.


	PR rtl-optimization/88233
	* common.opt (fsplit-wide-types-early): New option.
	* common/config/rs6000/rs6000-common.c
	(rs6000_option_optimization_table): Add OPT_fsplit_wide_types_early for
	OPT_LEVELS_ALL.
	* doc/invoke.texi (Optimization Options): Add -fsplit-wide-types-early.
	* lower-subreg.c (pass_lower_subreg2::gate): Add test for
	flag_split_wide_types_early.
	(pass_data_lower_subreg3): New.
	(pass_lower_subreg3): New.
	(make_pass_lower_subreg3): New.
	* passes.def (pass_lower_subreg2): Move after the loop passes.
	(pass_lower_subreg3): New, inserted where pass_lower_subreg2 was.
	* tree-pass.h (make_pass_lower_subreg2): Move up, to its new place in
	the pass pipeline; its previous place is taken by ...
	(make_pass_lower_subreg3): ... this.


git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@273240 138bc75d-0d04-0410-961f-82ee72b054a4