[Ada] Allow multiple units per file in GNATprove
For analysis tools that rely on information generated in ALI files, but
do not generate object files, the frontend did not generate the special
extension names like file~2.ali for unit 2 in the file. This is needed
to be able to analyze files with multiple units. Now fixed.

There is no impact on compilation.

2019-07-10  Yannick Moy  <moy@adacore.com>

gcc/ada/

	* osint-c.adb (Set_File_Name): Always add extension for multiple
	units per file mode.

git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@273339 138bc75d-0d04-0410-961f-82ee72b054a4