Relax vector_builder::elt sanity check
I'd made it a precondition of vector_builder::elt that the encoding
must have been fully populated and that all implicit elements are
therefore defined.  But for one of the AArch64 patches I'm working on,
it'd be convenient to be able to look back at previous elements while
building up the encoding.  This patch therefore makes the assert
specific to implicit elements only.

2019-07-12  Richard Sandiford  <richard.sandiford@arm.com>

gcc/
	* vector-builder.h (vector_builder::elt): Allow already-supplied
	elements to be read back before building is complete.


git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@273440 138bc75d-0d04-0410-961f-82ee72b054a4